+++ packages/colony-js-client/src/TokenLockingClient/index.js
@@ -51,7 +51,7 @@ export default class TokenLockingClient extends ContractClient {
     {
       contract: 'TokenLocking.sol',
       interface: 'ITokenLocking.sol',
-      version: 'glider-rc.1',
+      version: 'glider',
     },
   >;
 
@@ -69,7 +69,7 @@ export default class TokenLockingClient extends ContractClient {
     {
       contract: 'TokenLocking.sol',
       interface: 'ITokenLocking.sol',
-      version: 'glider-rc.1',
+      version: 'glider',
     },
   >;
 
@@ -88,7 +88,7 @@ export default class TokenLockingClient extends ContractClient {
     {
       contract: 'TokenLocking.sol',
       interface: 'ITokenLocking.sol',
-      version: 'glider-rc.1',
+      version: 'glider',
     },
   >;
 
@@ -106,7 +106,7 @@ export default class TokenLockingClient extends ContractClient {
     {
       contract: 'TokenLocking.sol',
       interface: 'ITokenLocking.sol',
-      version: 'glider-rc.1',
+      version: 'glider',
     },
   >;
 
@@ -123,7 +123,7 @@ export default class TokenLockingClient extends ContractClient {
     {
       contract: 'TokenLocking.sol',
       interface: 'ITokenLocking.sol',
-      version: 'glider-rc.1',
+      version: 'glider',
     },
   >;
 
@@ -143,7 +143,7 @@ export default class TokenLockingClient extends ContractClient {
     {
       contract: 'TokenLocking.sol',
       interface: 'ITokenLocking.sol',
-      version: 'glider-rc.1',
+      version: 'glider',
     },
   >;
 
@@ -162,7 +162,7 @@ export default class TokenLockingClient extends ContractClient {
     {
       contract: 'TokenLocking.sol',
       interface: 'ITokenLocking.sol',
-      version: 'glider-rc.1',
+      version: 'glider',
     },
   >;