+++ packages/colony-js-client/src/getNetworkClient.js
@@ -1,6 +1,8 @@
 /* @flow */
 /* eslint-disable import/no-cycle */
 
+import type { WalletObjectType } from '@colony/purser-core';
+
 import { providers } from 'ethers';
 import EthersAdapter from '@colony/colony-js-adapter-ethers';
 import NetworkLoader from '@colony/colony-js-contract-loader-network';
@@ -15,19 +17,11 @@ const getInfuraProvider = (network: string, infuraProjectId?: string) => {
   let host;
   switch (network) {
     case 'homestead':
+    case 'mainnet':
       host = 'mainnet.infura.io';
-      break;
-    case 'ropsten':
-      host = 'ropsten.infura.io';
-      break;
-    case 'rinkeby':
-      host = 'rinkeby.infura.io';
       break;
     case 'goerli':
       host = 'goerli.infura.io';
-      break;
-    case 'kovan':
-      host = 'kovan.infura.io';
       break;
     default:
       throw new Error(
@@ -52,7 +46,7 @@ const getInfuraProvider = (network: string, infuraProjectId?: string) => {
 // local testrpc network (TrufflePig must be installed and running).
 const getNetworkClient = async (
   network: string,
-  wallet: any,
+  wallet: WalletObjectType,
   infuraProjectId?: string,
 ) => {
   let loader;