+++ packages/colony-js-contract-loader-network/src/NetworkLoader.js
@@ -8,84 +8,61 @@ import ContractLoader, {
 import type { RequiredContractProps } from '@colony/colony-js-contract-loader';
 
 const NETWORKS = {
-  RINKEBY: 'rinkeby',
   GOERLI: 'goerli',
+  MAINNET: 'mainnet',
 };
 
-const DEFAULT_NETWORK = NETWORKS.GOERLI;
+const DEFAULT_NETWORK = NETWORKS.MAINNET;
 
 type Network = $Values<typeof NETWORKS>;
 
-const LATEST_VERSION = 1;
+const LATEST_VERSION = 2;
 
 const CONTRACTS_MANIFEST = {
-  // static: [],
   versioned: {
-    [NETWORKS.RINKEBY]: {
-      '1': ['Authority', 'EtherRouter', 'IColony', 'IColonyNetwork', 'Token'],
-      '2': ['Authority', 'EtherRouter', 'IColony', 'IColonyNetwork', 'Token'],
-      '3': [
-        'EtherRouter',
-        'IColony',
-        'IColonyNetwork',
-        'IMetaColony',
-        'IRecovery',
-        'Token',
-        'DSToken',
-        'ITokenLocking',
-        'OneTxPayment',
-      ],
-    },
-    [NETWORKS.GOERLI]: {
-      '1': [
-        'EtherRouter',
-        'IColony',
-        'IColonyNetwork',
-        'IMetaColony',
-        'IRecovery',
-        'ITokenLocking',
-        'OldRoles',
-        'OldRolesFactory',
-        'OneTxPayment',
-        'OneTxPaymentFactory',
-        'Token',
-        'TokenAuthority',
-      ],
-    },
+    '1': [
+      'EtherRouter',
+      'IColony',
+      'IColonyNetwork',
+      'IMetaColony',
+      'IRecovery',
+      'ITokenLocking',
+      'OldRoles',
+      'OldRolesFactory',
+      'OneTxPayment',
+      'OneTxPaymentFactory',
+      'Token',
+      'TokenAuthority',
+    ],
+    '2': [
+      'EtherRouter',
+      'IColony',
+      'IColonyNetwork',
+      'IMetaColony',
+      'IRecovery',
+      'ITokenLocking',
+      'OldRoles',
+      'OldRolesFactory',
+      'OneTxPayment',
+      'OneTxPaymentFactory',
+      'Token',
+      'TokenAuthority',
+    ],
   },
 };
 
-// const STATIC_CONTRACTS = CONTRACTS_MANIFEST.static.reduce(
-//   (contracts, contractName) =>
-//     Object.assign(contracts, {
-//       // eslint-disable-next-line global-require, import/no-dynamic-require
-//       [contractName]: require(`../contracts/static/${contractName}.json`),
-//     }),
-//   {},
-// );
-
 const VERSIONED_CONTRACTS = Object.entries(CONTRACTS_MANIFEST.versioned).reduce(
-  (networks, [network, versions]) =>
-    Object.assign(
-      networks,
-      {
-        [network]: Object.entries(versions).reduce(
-          (versionedContracts, [version, contractNames]) =>
-            Object.assign(versionedContracts, {
-              [version]: contractNames.reduce(
-                (contracts, contractName) =>
-                  Object.assign(contracts, {
-                    // eslint-disable-next-line global-require, import/no-dynamic-require, max-len
-                    [contractName]: require(`../contracts/versioned/${network}-v${version}/${contractName}.json`),
-                  }),
-                {},
-              ),
-            }),
-          {},
-        ),
-      },
-      {},
-    ),
+  (versionedContracts, [version, contractNames]) =>
+    Object.assign(versionedContracts, {
+      [version]: contractNames.reduce(
+        (contracts, contractName) =>
+          Object.assign(contracts, {
+            // eslint-disable-next-line global-require, import/no-dynamic-require, max-len
+            [contractName]: require(`../contracts/versioned/v${version}/${contractName}.json`),
+          }),
+        {},
+      ),
+    }),
   {},
 );
 
@@ -106,24 +83,15 @@ class NetworkLoader extends ContractLoader {
     assert(!!contractName, 'A `contractName` option must be provided');
     assert(!!version, 'A valid `version` option must be provided');
 
-    // if (STATIC_CONTRACTS[contractName]) {
-    //   return this._transform(
-    //     STATIC_CONTRACTS[contractName],
-    //     networkQuery,
-    //     requiredProps,
-    //   );
-    // }
-
     const contract =
-      VERSIONED_CONTRACTS[network] &&
-      VERSIONED_CONTRACTS[network][version] &&
-      VERSIONED_CONTRACTS[network][version][contractName];
+      VERSIONED_CONTRACTS[version] &&
+      VERSIONED_CONTRACTS[version][contractName];
 
     if (contract) return this._transform(contract, networkQuery, requiredProps);
 
     throw new Error(
       // eslint-disable-next-line max-len
-      `Contract ${contractName} with version ${version} not found in ${network}`,
+      `Contract ${contractName} with version ${version} not found on ${network}`,
     );
   }
 }